Split PlanViewerControl.axaml.cs into partial classes
Move-only refactor; no behavior changes. PlanViewerControl.axaml.cs
(4,497 lines) split into 7 partials:

  Rendering    (550) - RenderStatement/Nodes/Edges + edge tooltips
  Properties   (1860) - ShowPropertiesPanel + all property/runtime/wait panels
  Tooltips     (278) - BuildNodeTooltipContent + helpers
  Interaction  (327) - Node_Click, Select, zoom, pointer events, save
  Statements   (222) - statements grid panel
  Minimap      (502) - minimap render/drag/resize/navigation
  Schema       (347) - context-menu schema lookup + index/column formatting

Main file (PlanViewerControl.axaml.cs) now 504 lines — fields, brushes,
constructor, public API (LoadPlan/Clear/NavigateToNode), connection
state, and PlanConnect/PlanDatabase event handlers.

Build clean: 0 errors, 0 warnings on PlanViewer.sln.
